Fix Window rendering  and ScreenspaceRect coords
Previously, the Rect of a Window would represent the size of the Windows Screen, but as it's the Window and not the Screen that owns the Rect it makes more sense for the Rect to represent the Windows size instead. The Screen, as a result, is now two smaller on each axis than it was previously. This also enables the addition of borderless Windows, which will have Screens of the same size as the Window.
ScreenspaceRect coordinates have been fixed too, so that the behaviour of a ScreenspaceRect is to spec. Closes #4.