Merge prerequisite function call into getKey
Previously, the updateKeyBuffer() method had to be called manually to move keys from stdin to the key buffer. Since the only point of access for the keybuffer is getKey, it makes sense for getKey to call updateKeyBuffer itself. This means that the user never has to remember to call it manually before they call getKey.